split store.py into a module with multiple small files
