+++ .flake8
@@ -1,0 +1,3 @@
+[flake8]
+per-file-ignores = */__init__.py:F401
+max-line-length = 120
