+++ packages/modules/common/store/_api.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+from typing import Generic, TypeVar
+
+T = TypeVar("T")
+
+
+class ValueStore(Generic[T]):
+    @abstractmethod
+    def set(self, state: T) -> None:
+        pass
